survey v2 添加用户和角色模块

+++ src/com/dave/controller/PaperController.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
@@ -58,6 +59,9 @@ public class PaperController {
     @RequestMapping("doSelectQues")
     @ResponseBody
     public JsonResult doSelectQues(Integer... quesIds) {
+    	if(StringUtils.isEmpty(quesIds) || quesIds.length > 0) {
+    		return new JsonResult("问题ID不能为空");
+    	}
     	List<QuesInfo> list = paperService.selectQuesByIds(quesIds);
     	return new JsonResult(list);
     }
@@ -70,7 +74,31 @@ public class PaperController {
     @ResponseBody
     public JsonResult doAddPaper(PaperInfo paperInfo) {
     	if(paperInfo == null) {
-    		return new JsonResult("参数对象为空");
+    		return new JsonResult("问卷对象不能为空");
+    	}
+    	if(StringUtils.isEmpty(paperInfo.getPaperName())) {
+    		return new JsonResult("问卷名称不能为空");
+    	}
+    	if(StringUtils.isEmpty(paperInfo.getPaperTitle())) {
+    		return new JsonResult("问卷标题不能为空");
+    	}
+    	if(StringUtils.isEmpty(paperInfo.getPaperLanguage())) {
+    		return new JsonResult("问卷语言不能为空");
+    	}
+    	if(StringUtils.isEmpty(paperInfo.getGreet())) {
+    		return new JsonResult("问卷欢迎内容不能为空");
+    	}
+    	if(StringUtils.isEmpty(paperInfo.getPaperType())) {
+    		return new JsonResult("问卷类型不能为空");
+    	}
+    	if(StringUtils.isEmpty(paperInfo.getThank())) {
+    		return new JsonResult("问卷感谢内容不能为空");
+    	}
+    	if(StringUtils.isEmpty(paperInfo.getQuesIds()) || paperInfo.getQuesIds().length > 0) {
+    		return new JsonResult("问题ID集不能为空");
+    	}
+    	if(StringUtils.isEmpty(paperInfo.getQuesNum()) || paperInfo.getQuesNum().length > 0) {
+    		return new JsonResult("问题序号不能为空");
     	}
     	int rowName = paperService.checkoutPaperName(paperInfo.getPaperName(), paperInfo.getPaperLanguage(), paperInfo.getPaperId());
     	if(rowName > 0) {
@@ -89,7 +117,10 @@ public class PaperController {
      */
     @RequestMapping("doFindPaperList")
     @ResponseBody
-    public JsonResult doFindPaperList(int pageCurrent, String paperName) {
+    public JsonResult doFindPaperList(Integer pageCurrent, String paperName) {
+    	if(pageCurrent == null || pageCurrent <= 0) {
+    		return new JsonResult("当前页面数不能为空");
+    	}
     	PageObject<Paper> paperList = paperService.findPaperList(pageCurrent, paperName);
     	return new JsonResult(paperList);
     }
@@ -101,6 +132,9 @@ public class PaperController {
     @RequestMapping("doDeletePaper")
     @ResponseBody
     public JsonResult doDeletePaper(Integer... paperIds) {
+    	if(StringUtils.isEmpty(paperIds) || paperIds.length > 0) {
+    		return new JsonResult("问卷ID集不能为空");
+    	}
     	int row = paperService.deletePaper(paperIds);
     	if(row > 0) {
     		return new JsonResult("Delete Succeed!", row); 		
@@ -115,6 +149,12 @@ public class PaperController {
     @RequestMapping("doUpdateStatus")
     @ResponseBody
     public JsonResult doUpdateStatus(Paper paper) {
+    	if(paper == null) {
+    		return new JsonResult("问卷对象不能为空");
+    	}
+    	if(StringUtils.isEmpty(paper.getStatus())) {
+    		return new JsonResult("问卷状态不能为空");
+    	}
     	int row = paperService.updateStatus(paper);
     	if(row > 0) {
     		return new JsonResult("Update Succeed!", row); 		
@@ -128,7 +168,10 @@ public class PaperController {
      */
     @RequestMapping("doGetPaperQues")
     @ResponseBody
-    public JsonResult doGetPaperQues(int paperId) {
+    public JsonResult doGetPaperQues(Integer paperId) {
+    	if(StringUtils.isEmpty(paperId)) {
+    		return new JsonResult("问卷ID不能为空");
+    	}
     	PaperInfo paperInfo = paperService.getPaperQues(paperId);
     	return new JsonResult(paperInfo);
     }
@@ -139,7 +182,10 @@ public class PaperController {
      */
     @RequestMapping("doGetPaperQuesOption")
     @ResponseBody
-    public JsonResult doGetPaperQuesOption(int paperId) {
+    public JsonResult doGetPaperQuesOption(Integer paperId) {
+    	if(StringUtils.isEmpty(paperId)) {
+    		return new JsonResult("问卷ID不能为空");
+    	}
     	List<PaperQuesOption> list = paperService.getPaperQuesOption(paperId);
     	return new JsonResult(list);
     }
@@ -152,7 +198,34 @@ public class PaperController {
     @ResponseBody
     public JsonResult doUpdatePaper(PaperInfo paperInfo) {
     	if(paperInfo == null) {
-    		return new JsonResult("参数对象为空");
+    		return new JsonResult("问卷对象不能为空");
+    	}
+    	if(StringUtils.isEmpty(paperInfo.getPaperId())) {
+    		return new JsonResult("问卷ID不能为空");
+    	}
+    	if(StringUtils.isEmpty(paperInfo.getPaperName())) {
+    		return new JsonResult("问卷名称不能为空");
+    	}
+    	if(StringUtils.isEmpty(paperInfo.getPaperTitle())) {
+    		return new JsonResult("问卷标题不能为空");
+    	}
+    	if(StringUtils.isEmpty(paperInfo.getPaperLanguage())) {
+    		return new JsonResult("问卷语言不能为空");
+    	}
+    	if(StringUtils.isEmpty(paperInfo.getGreet())) {
+    		return new JsonResult("问卷欢迎内容不能为空");
+    	}
+    	if(StringUtils.isEmpty(paperInfo.getPaperType())) {
+    		return new JsonResult("问卷类型不能为空");
+    	}
+    	if(StringUtils.isEmpty(paperInfo.getThank())) {
+    		return new JsonResult("问卷感谢内容不能为空");
+    	}
+    	if(StringUtils.isEmpty(paperInfo.getQuesIds()) || paperInfo.getQuesIds().length > 0) {
+    		return new JsonResult("问题ID集不能为空");
+    	}
+    	if(StringUtils.isEmpty(paperInfo.getQuesNum()) || paperInfo.getQuesNum().length > 0) {
+    		return new JsonResult("问题序号不能为空");
     	}
     	int rowName = paperService.checkoutPaperName(paperInfo.getPaperName(), paperInfo.getPaperLanguage(), paperInfo.getPaperId());
     	if(rowName > 0) {
